Move interface.ts to local-queries/results-view.ts
