+++ extensions/ql-vscode/src/local-queries/index.ts
@@ -2,4 +2,5 @@ export * from "./local-queries";
 export * from "./local-query-run";
 export * from "./quick-eval-code-lens-provider";
 export * from "./quick-query";
+export * from "./results-view";
 export { WebviewReveal } from "./webview";
